refactor(import-network-panel): move text-panel to a separate file
